update fxmanifest/config and move files
